pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plc-telemetry"
version = "1.0.0"
description = "Python-first PLC telemetry platform MVP"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Codex" }
]
dependencies = [
  "matplotlib>=3.7",
  "polars>=1.8",
  "pyads>=3.4",
  "pyarrow>=14.0",
  "PySide6>=6.5",
  "pyqtgraph>=0.13",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
plc-telemetry = "plc_telemetry.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
